# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mind The Wait API",
        "description": "Mobile API for iOS app - Transit headway monitoring and predictions",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost",
            "description": "Local development"
        },
        {
            "url": "https://mindthewait.ca",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/v1/stops": {
            "get": {
                "tags": [
                    "Stops"
                ],
                "summary": "List stops",
                "description": "Returns all stops or stops for a specific route",
                "operationId": "get_api_v1_stops_list",
                "parameters": [
                    {
                        "name": "route_id",
                        "in": "query",
                        "description": "Filter stops by GTFS route ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of stops",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stops": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "4000"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "19th Street / 2nd Avenue"
                                                    },
                                                    "lat": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 52.124263
                                                    },
                                                    "lon": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": -106.666571
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "timestamp": {
                                            "type": "integer",
                                            "example": 1760939266
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/stops/{gtfsId}/predictions": {
            "get": {
                "tags": [
                    "Stops"
                ],
                "summary": "Get stop predictions",
                "description": "Returns next vehicle arrivals for a specific stop",
                "operationId": "get_api_v1_stop_predictions",
                "parameters": [
                    {
                        "name": "gtfsId",
                        "in": "path",
                        "description": "Stop GTFS ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of predictions to return (default: 5)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "route_id",
                        "in": "query",
                        "description": "Filter predictions by route ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arrival predictions for stop",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stop": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "4000"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "19th Street / 2nd Avenue"
                                                },
                                                "lat": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 52.124263
                                                },
                                                "lon": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": -106.666571
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "predictions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "route_id": {
                                                        "type": "string",
                                                        "example": "14514"
                                                    },
                                                    "route_short_name": {
                                                        "type": "string",
                                                        "example": "16"
                                                    },
                                                    "arrival_time": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-10-19T12:30:00Z"
                                                    },
                                                    "delay_sec": {
                                                        "type": "integer",
                                                        "example": 120
                                                    },
                                                    "confidence": {
                                                        "type": "string",
                                                        "enum": [
                                                            "HIGH",
                                                            "MEDIUM",
                                                            "LOW"
                                                        ],
                                                        "example": "HIGH"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "timestamp": {
                                            "type": "integer",
                                            "example": 1760939266
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Stop not found"
                    }
                }
            }
        },
...
        "/api/stops/{stopId}/predictions": {
            "get": {
                "operationId": "get_stop_predictions",
                "parameters": [
                    {
                        "name": "stopId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
...
}